Struct structured_data::structured_data::FixedAttributes

pub struct FixedAttributes {
    pub type_tag: u64,
    pub id: NameType,
    pub max_versions: u64,
    pub min_retained_count: u8,
    pub data: Vec<u8>,
}

Attributes of the Data which can never change once initially set. These define the identity, type and some of the rules the network will employ for handling the Data. It can also hold arbitrary data which will likely be meaningless to the network.

Fields

type_tag

Identifier of the Data type.

id

Identity of the piece of Data.

max_versions

Maximum number of versions allowed.

min_retained_count

Number of versions to retain when archiving a "full" piece of Data (minimum value of 1).

data

Arbitrary, immutable, Data-wide information. May be empty.