Struct structured_data::structured_data::MutableAttributes

pub struct MutableAttributes {
    pub owner_keys: Vec<KeyAndWeight>,
    pub min_weight_for_consensus: u64,
    pub expiry_date: Tm,
    pub data: Vec<u8>,
}

Attributes of the Data which can be changed via a properly-authorised request to the network. These define the current owner's public keys, further rules the network will employ for handling the Data and also arbitrary data which will likely be meaningless to the network.

Fields

owner_keys

Current owner or owners' public keys. Cannot be empty.

min_weight_for_consensus

Minimum total weight of signatories' keys to allow a mutation of the piece of Data (at least one signature will be required regardless of this minimum).

expiry_date

Coarse-grained expiry date around which time the piece of Data will be removed from the network.

data

Arbitrary, mutable, Data-wide information. May be empty.