Struct structured_data::structured_data::Version

pub struct Version {
    pub index: u64,
    pub data: Vec<u8>,
}

A representation of a single version. The index allows provision of strict total ordering of the Versions. It can also hold arbitrary data specific to that particular Version, e.g. encrypted content or the name of a piece of "Immutable Data".

Fields

index

Sequential number to provide strict total order of versions.

data

Arbitrary, version-specific information. May be empty.